requirements.txt
pandas>=2.0.0
pandera>=0.17.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
pytest>=7.0.0
pytest-cov>=4.0.0
